Type correctness fixes for some tables

+++ SF3Lib/Models/Tables/UnknownUInt16Table.cs
@@ -5,7 +5,7 @@ using SF3.Models.Structs;
 
 namespace SF3.Models.Tables {
     public class UnknownUInt16Table : TerminatedTable<UnknownUInt16Struct> {
-        protected UnknownUInt16Table(IByteData data, string name, int address, int? count, int? readUntil)
+        protected UnknownUInt16Table(IByteData data, string name, int address, int? count, ushort? readUntil)
         : base(data, name, address, readUntil.HasValue ? 2 : 0, count) {
             if (!count.HasValue && !readUntil.HasValue)
                 throw new ArgumentNullException(nameof(count) + ", " + nameof(readUntil));
@@ -14,7 +14,7 @@ namespace SF3.Models.Tables {
             FormatString = "X" + MaxSize.ToString().Length;
         }
 
-        public static UnknownUInt16Table Create(IByteData data, string name, int address, int? count, int? readUntil)
+        public static UnknownUInt16Table Create(IByteData data, string name, int address, int? count, ushort? readUntil)
             => Create(() => new UnknownUInt16Table(data, name, address, count, readUntil));
 
         private bool ContinueReadingPred(Dictionary<int, UnknownUInt16Struct> currentRows, UnknownUInt16Struct newModel)
@@ -25,7 +25,7 @@ namespace SF3.Models.Tables {
             return Load((id, address) => new UnknownUInt16Struct(Data, id, "Unknown UInt16 " + id.ToString(FormatString), address), pred, false);
         }
 
-        public int? ReadUntil { get; }
+        public ushort? ReadUntil { get; }
         private string FormatString { get; }
     }
 }
